feat: add RPC mode support for transaction bridging
Implement dual-mode operation supporting both P2P (devp2p) and RPC (JSON-RPC) protocols for bridging mempool transactions between execution layer nodes.

Changes:
- Add mode configuration field to switch between P2P and RPC modes
- Implement RPC source coordinator with WebSocket subscriptions to newPendingTransactions
- Implement RPC target coordinator with eth_sendRawTransaction support
- Refactor bridge to use interface-based coordinator pattern
- Reorganize P2P coordinators into dedicated subdirectories
- Add comprehensive documentation and example configs for both modes
- Update README with feature overview and configuration examples

The implementation maintains full backward compatibility with existing P2P mode while enabling RPC-based transaction bridging for nodes that prefer JSON-RPC over devp2p.
