--- example_config.yaml
@@ -1,24 +1,0 @@
-logging: "info" # panic,fatal,warn,info,debug,trace
-metricsAddr: ":9090"
-
-# nodes to source mempool transactions from
-source:
-  retryInterval: 30s
-  # optional filter transactions by to address
-  # transactionFilters:
-  #   to:
-  #     - 0x0000000000000000000000000000000000000000
-  #     - 0x1111111111111111111111111111111111111111
-  #   from:
-  #     - 0x2222222222222222222222222222222222222222
-  #     - 0x3333333333333333333333333333333333333333
-  nodeRecords:
-    - enr:-IS4QHCYrYZbAKWCBRlAy5zzaDZXJBGkcnh4MHcBFZntXNFrdvJjX04jRzjzCBOonrkTfj499SZuOh8R33Ls8RRcy5wBgmlkgnY0gmlwhH8AAAGJc2VjcDI1NmsxoQPKY0yuDUmstAHYpMa2_oxVtw0RW_QAdpzBQA8yWM0xOIN1ZHCCdl8
-    - enode://dd47aff4da11c82496fec9253a1d2dac8c4abdbf93ff7458cf7965447355c265907b41dbd8c4b1dded6ae28b5a4d2d6b4d3fdad2950e5b6ef56169958b80b450@127.0.0.1:30301
-
-# nodes to target transactions to
-target:
-  retryInterval: 30s
-  nodeRecords:
-    - enr:-IS4QHCYrYZbAKWCBRlAy5zzaDZXJBGkcnh4MHcBFZntXNFrdvJjX04jRzjzCBOonrkTfj499SZuOh8R33Ls8RRcy5wBgmlkgnY0gmlwhH8AAAGJc2VjcDI1NmsxoQPKY0yuDUmstAHYpMa2_oxVtw0RW_QAdpzBQA8yWM0xOIN1ZHCCdl8
-    - enode://dd47aff4da11c82496fec9253a1d2dac8c4abdbf93ff7458cf7965447355c265907b41dbd8c4b1dded6ae28b5a4d2d6b4d3fdad2950e5b6ef56169958b80b450@127.0.0.1:30301
