+++ example_config_mixed.yaml
@@ -1,0 +1,44 @@
+# Example configuration showing both P2P and RPC configurations
+# Note: Only one mode can be active at a time based on the "mode" setting
+
+logging: "info" # panic,fatal,warn,info,debug,trace
+metricsAddr: ":9090"
+
+# Mode selection: "p2p" or "rpc"
+# Change this to switch between P2P and RPC modes
+mode: "p2p"
+
+# Source configuration
+source:
+  retryInterval: 30s
+
+  # P2P mode: ENR/enode records for devp2p connections
+  nodeRecords:
+    - enr:-IS4QHCYrYZbAKWCBRlAy5zzaDZXJBGkcnh4MHcBFZntXNFrdvJjX04jRzjzCBOonrkTfj499SZuOh8R33Ls8RRcy5wBgmlkgnY0gmlwhH8AAAGJc2VjcDI1NmsxoQPKY0yuDUmstAHYpMa2_oxVtw0RW_QAdpzBQA8yWM0xOIN1ZHCCdl8
+    - enode://dd47aff4da11c82496fec9253a1d2dac8c4abdbf93ff7458cf7965447355c265907b41dbd8c4b1dded6ae28b5a4d2d6b4d3fdad2950e5b6ef56169958b80b450@127.0.0.1:30301
+
+  # RPC mode: WebSocket endpoints (ws:// or wss:// required)
+  rpcEndpoints:
+    - ws://localhost:8546
+    - ws://node2.example.com:8546
+
+  # optional filter transactions by to/from address (works in both modes)
+  # transactionFilters:
+  #   to:
+  #     - 0x0000000000000000000000000000000000000000
+  #   from:
+  #     - 0x2222222222222222222222222222222222222222
+
+# Target configuration
+target:
+  retryInterval: 30s
+
+  # P2P mode: ENR/enode records for devp2p connections
+  nodeRecords:
+    - enr:-IS4QHCYrYZbAKWCBRlAy5zzaDZXJBGkcnh4MHcBFZntXNFrdvJjX04jRzjzCBOonrkTfj499SZuOh8R33Ls8RRcy5wBgmlkgnY0gmlwhH8AAAGJc2VjcDI1NmsxoQPKY0yuDUmstAHYpMa2_oxVtw0RW_QAdpzBQA8yWM0xOIN1ZHCCdl8
+    - enode://dd47aff4da11c82496fec9253a1d2dac8c4abdbf93ff7458cf7965447355c265907b41dbd8c4b1dded6ae28b5a4d2d6b4d3fdad2950e5b6ef56169958b80b450@127.0.0.1:30301
+
+  # RPC mode: HTTP or WebSocket endpoints
+  rpcEndpoints:
+    - http://localhost:8545
+    - http://node2.example.com:8545
