+++ example_config_p2p.yaml
@@ -1,0 +1,32 @@
+logging: "info" # panic,fatal,warn,info,debug,trace
+metricsAddr: ":9090"
+
+# Mode selection: "p2p" or "rpc"
+# When mode is "p2p", ONLY nodeRecords are used (rpcEndpoints are ignored)
+# When mode is "rpc", ONLY rpcEndpoints are used (nodeRecords are ignored)
+mode: "p2p"
+
+# Source: Where to listen for mempool transactions
+source:
+  retryInterval: 30s
+
+  # P2P mode: ENR/enode records for devp2p connections
+  nodeRecords:
+    - enr:-IS4QHCYrYZbAKWCBRlAy5zzaDZXJBGkcnh4MHcBFZntXNFrdvJjX04jRzjzCBOonrkTfj499SZuOh8R33Ls8RRcy5wBgmlkgnY0gmlwhH8AAAGJc2VjcDI1NmsxoQPKY0yuDUmstAHYpMa2_oxVtw0RW_QAdpzBQA8yWM0xOIN1ZHCCdl8
+    - enode://dd47aff4da11c82496fec9253a1d2dac8c4abdbf93ff7458cf7965447355c265907b41dbd8c4b1dded6ae28b5a4d2d6b4d3fdad2950e5b6ef56169958b80b450@127.0.0.1:30301
+
+  # Optional: Filter transactions by to/from addresses
+  # transactionFilters:
+  #   to:
+  #     - 0x0000000000000000000000000000000000000000
+  #   from:
+  #     - 0x2222222222222222222222222222222222222222
+
+# Target: Where to send transactions
+target:
+  retryInterval: 30s
+
+  # P2P mode: ENR/enode records for devp2p connections
+  nodeRecords:
+    - enr:-IS4QHCYrYZbAKWCBRlAy5zzaDZXJBGkcnh4MHcBFZntXNFrdvJjX04jRzjzCBOonrkTfj499SZuOh8R33Ls8RRcy5wBgmlkgnY0gmlwhH8AAAGJc2VjcDI1NmsxoQPKY0yuDUmstAHYpMa2_oxVtw0RW_QAdpzBQA8yWM0xOIN1ZHCCdl8
+    - enode://dd47aff4da11c82496fec9253a1d2dac8c4abdbf93ff7458cf7965447355c265907b41dbd8c4b1dded6ae28b5a4d2d6b4d3fdad2950e5b6ef56169958b80b450@127.0.0.1:30301
