+++ pkg/bridge/bridge.go
@@ -2,6 +2,7 @@ package bridge
 
 import (
 	"context"
+	"fmt"
 	"net/http"
 	"os"
 	"os/signal"
@@ -10,17 +11,34 @@ import (
 
 	"github.com/ethpandaops/ethcore/pkg/execution/mimicry"
 	"github.com/ethpandaops/mempool-bridge/pkg/bridge/source"
+	sourcep2p "github.com/ethpandaops/mempool-bridge/pkg/bridge/source/p2p"
+	sourcerpc "github.com/ethpandaops/mempool-bridge/pkg/bridge/source/rpc"
 	"github.com/ethpandaops/mempool-bridge/pkg/bridge/target"
+	targetp2p "github.com/ethpandaops/mempool-bridge/pkg/bridge/target/p2p"
+	targetrpc "github.com/ethpandaops/mempool-bridge/pkg/bridge/target/rpc"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/sirupsen/logrus"
 )
+
+// SourceCoordinator defines the interface for source coordinators
+type SourceCoordinator interface {
+	Start(ctx context.Context) error
+	Stop(ctx context.Context) error
+}
+
+// TargetCoordinator defines the interface for target coordinators
+type TargetCoordinator interface {
+	Start(ctx context.Context) error
+	Stop(ctx context.Context) error
+	SendTransactionsToPeers(ctx context.Context, transactions *mimicry.Transactions) error
+}
 
 type Bridge struct {
 	log *logrus.Logger
 	Cfg Config
 
-	s *source.Coordinator
-	t *target.Coordinator
+	s SourceCoordinator
+	t TargetCoordinator
 
 	metrics *Metrics
 }
@@ -30,26 +48,55 @@ func New(log *logrus.Logger, conf *Config) *Bridge {
 		log.Fatalf("invalid config: %s", err)
 	}
 
-	t, err := target.NewCoordinator(&conf.Target, log)
+	b := &Bridge{
+		Cfg:     *conf,
+		log:     log,
+		metrics: NewMetrics("mempool_bridge"),
+	}
+
+	// Create target coordinator based on mode
+	t, err := createTargetCoordinator(conf.Mode, &conf.Target, log)
 	if err != nil {
 		log.Fatalf("failed to create target: %s", err)
 	}
 
-	b := &Bridge{
-		Cfg:     *conf,
-		log:     log,
-		t:       t,
-		metrics: NewMetrics("mempool_bridge"),
-	}
+	b.t = t
 
-	s, err := source.NewCoordinator(&conf.Source, b.broadcast, log)
+	// Create source coordinator based on mode
+	s, err := createSourceCoordinator(conf.Mode, &conf.Source, b.broadcast, log)
 	if err != nil {
 		log.Fatalf("failed to create source: %s", err)
 	}
 
 	b.s = s
 
+	log.WithField("mode", conf.Mode).Info("initialized bridge")
+
 	return b
+}
+
+// createTargetCoordinator creates a target coordinator based on the mode
+func createTargetCoordinator(mode Mode, config *target.Config, log *logrus.Logger) (TargetCoordinator, error) {
+	switch mode {
+	case ModeP2P:
+		return targetp2p.NewCoordinator(config, log)
+	case ModeRPC:
+		return targetrpc.NewCoordinator(config, log)
+	default:
+		return nil, fmt.Errorf("unknown mode: %s", mode)
+	}
+}
+
+// createSourceCoordinator creates a source coordinator based on the mode
+func createSourceCoordinator(mode Mode, config *source.Config, broadcast func(ctx context.Context, transactions *mimicry.Transactions) error, log *logrus.Logger) (SourceCoordinator, error) {
+	switch mode {
+	case ModeP2P:
+		return sourcep2p.NewCoordinator(config, broadcast, log)
+	case ModeRPC:
+		return sourcerpc.NewCoordinator(config, broadcast, log)
+	default:
+		return nil, fmt.Errorf("unknown mode: %s", mode)
+	}
 }
 
 func (b *Bridge) Start(ctx context.Context) error {
